requirements.txt
Flask==2.0.1
PyYAML==6.0.1
rendercv>=0.5.0
Werkzeug==2.0.1
python-dotenv==0.19.0
boto3==1.26.137
PyPDF2==3.0.1
requests==2.28.2
python-magic==0.4.27
flask-cors==4.0.0
pdfminer.six==20221105
nltk>=3.6.2
pytesseract==0.3.10
Pillow>=10.4.0
pdf2image==1.16.3
pymongo==4.6.1
dnspython==2.5.0
numpy>=1.19.0
scipy>=1.6.0
joblib>=1.0.0
pandas>=1.3.0 